requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.39
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
email-validator==2.2.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
alembic==1.14.1
redis==5.2.1
